remove HTML from package description
The npm website now allows and renders some HTML tags in package descriptions, like `<strong>` and `<a>`. The full list is here: https://www.npmjs.com/package/sanitize-html#how-to-use

`<input>` is not on the whitelist though, so it'll get stripped out. This change works around that. by using a plain English description.