Add support for accented characters
Library from this SO answer : http://stackoverflow.com/a/9667817/2726340
I added a normalize() call to be sure about the encoding. On my mac, it appears that my filenames were in NFD, the library was in UTF8, so nothing was matched.

Now it seems to be working fine. Don't hesitate to make changements if necessary of course.

PR from issue #32