communicate S3 website URL to the client
In case S3 bucket is configured in static website hosting mode,
it is convenient to tell the client the public URL of newly created object
to access it directly (instead of going through s3router redirection).

Also, providing the function instead, it is possible to generate custom URL
for example, if there a CDN or custom domain name configured.